requirements.txt
Flask>=2.2
PyYAML>=6.0.1
requests>=2.25
pycryptodome>=3.19.0
pytest>=7.0
pytest-flask>=1.3.0
